requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
SQLAlchemy==2.0.40
requests==2.32.3
pydantic==2.11.3
python-dotenv==1.0.1